Introduce QUARANTINED state
Detailed proposals are at go/quarantined-when-not-in-use and go/disabled-until-launched-u

+block broadcasts
+force stop packages on suspension+quarantine

Test: Used command `pm suspend-quarantine <app_name>`.
Tested locally on device for basic functionality of putting apps
in and out of this state, and verifying some bindings and behaviors like
apps continuing to appear on launcher. More thorough testing to follow.
Test: atest DevicePolicyManagerServiceMigrationTest SuspendPackageHelperTest PackageManagerSettingsTests
Bug: 269127435

+++ services/core/java/com/android/server/apphibernation/AppHibernationService.java
@@ -97,7 +97,7 @@ import java.util.concurrent.ScheduledExecutorService;
  */
 public final class AppHibernationService extends SystemService {
     private static final String TAG = "AppHibernationService";
-    private static final int PACKAGE_MATCH_FLAGS =
+    private static final long PACKAGE_MATCH_FLAGS =
             PackageManager.MATCH_DIRECT_BOOT_AWARE
                     | PackageManager.MATCH_DIRECT_BOOT_UNAWARE
                     | PackageManager.MATCH_UNINSTALLED_PACKAGES
